converted some .js to .vue
