+++ .eslintrc
@@ -7,12 +7,11 @@
   "extends": "airbnb-base",
   "globals": {
     "Bloodhound": false,
-    "layout_resizer": false,
-    "set_typeahead": false,
-    "open_close_icon": false
+    "layout_resizer": false
   },
   "plugins": [
-    "import"
+    "import",
+    "html"
   ],
   "settings": {
     "import/resolver": {
@@ -27,9 +26,16 @@
     "prefer-arrow-callback": 0,
     "no-var": 0,
     "prefer-template": 0,
-    "no-unused-vars": [2, {"args": "all", "argsIgnorePattern": "^_"}],
+    "no-unused-vars": [2, {
+      "args": "all",
+      "argsIgnorePattern": "^_"
+    }],
     "object-shorthand": 0,
     "no-plusplus": 0,
-    "no-void": 0
+    "no-void": 0,
+    "import/extensions": ["error", "always", {
+      "js": "never",
+      "vue": "never"
+    }]
   }
 }
