--- app/assets/javascripts/modules/repositories/components/tags-not-loaded.js
@@ -1,7 +1,0 @@
-export default {
-  template: `
-    <div class="text-center">
-      Unable to fetch tags data.
-    </div>
-  `,
-};
