+++ app/assets/javascripts/modules/repositories/components/tags-not-loaded.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="text-center">
+    Unable to fetch tags data.
+  </div>
+</template>
