--- app/assets/javascripts/shared/components/loading-icon.js
@@ -1,20 +1,0 @@
-export default {
-  props: {
-    size: {
-      type: [String, Number],
-      default: 2,
-    },
-  },
-
-  computed: {
-    cssClass() {
-      return `fa-${this.size}x`;
-    },
-  },
-
-  template: `
-    <div class="text-center">
-      <i class="fa fa-spin fa-spinner" :class="cssClass"></i>
-    </div>
-  `,
-};
