+++ package.json
@@ -11,6 +11,7 @@
     "eslint": "^3.16.1",
     "eslint-config-airbnb-base": "^11.1.0",
     "eslint-import-resolver-webpack": "^0.8.1",
+    "eslint-plugin-html": "^2.0.3",
     "eslint-plugin-import": "^2.2.0",
     "webpack-dev-server": "^2.3.0"
   },
@@ -24,6 +25,7 @@
     "vue": "^2.3.3",
     "vue-loader": "^12.0.3",
     "vue-resource": "^1.3.1",
+    "vue-template-compiler": "^2.3.3",
     "webpack": "^2.2.1"
   },
   "babel": {
